RF: BF: ENH: Move map_colors and strip_prefix to research.data; fix map_colors bugs and work with None/NaN
